requirements.txt
# Core LLM & API
groq>=0.4.2
ollama>=0.1.8
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0

# Knowledge Graph
neo4j>=5.14.0
networkx>=3.2.1
pyvis>=0.3.2

# Vector DB & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.3.1

# NLP
spacy>=3.7.2
# Run after install: python -m spacy download en_core_web_sm

# Document Loading
pypdf>=3.17.4
python-docx>=1.1.0
unstructured>=0.11.0

# UI
streamlit>=1.31.0
streamlit-agraph>=0.0.45

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
tenacity>=8.2.3
loguru>=0.7.2
numpy>=1.26.0
pandas>=2.1.4
tqdm>=4.66.1
httpx>=0.26.0

# Caching
diskcache>=5.6.3

# Testing
pytest>=7.4.4
